refactor(native): isolate msvc toolchain policy
